refactor: extract mailbox.py with Mailbox class from mail _legacy
